Added demos of the MPU-6050 talking over I2C
I used an arduino to communicate with it. One demo is some basic stuf I
wrote using the Wire library and just shows one of the accelerometer
axes. The other is taken from online and demonstrates all 6 axes and
more. Also, I added the documentation for the MPU-6050 (and MPU-6000).